consolidates and groups recipe display, for clarity
