+++ src/features/RecipeDisplay/PlannedBucketController.tsx
@@ -1,59 +1,13 @@
 import useFluxStore from "data/useFluxStore";
-import getBucketLabel from "features/Planner/components/getBucketLabel";
 import TaskStore from "features/Planner/data/TaskStore";
 import LibraryStore from "features/RecipeLibrary/data/LibraryStore";
 import React from "react";
 import LoadObject from "util/LoadObject";
 import LoadingIndicator from "views/common/LoadingIndicator";
 import RecipeDetail from "./components/RecipeDetail";
-import {
-    buildFullRecipeLO as buildSingleTaskRecipeLO,
-    RecipeFromTask,
-} from "./PlannedRecipeController";
 import { RouteComponentProps } from "react-router";
-import { useLoadedPlan } from "./hooks/useLoadedPlan";
-
-export const buildFullRecipeLO = (planId: number, bucketId: number): LoadObject<RecipeFromTask> => {
-    const plan = TaskStore.getTaskLO(planId);
-    if (!plan.hasValue()) {
-        // no value means value's type is irrelevant
-        return plan as LoadObject<any>;
-    }
-    const bucket = plan.getValueEnforcing()
-        .buckets
-        .find(b => b.id === bucketId);
-    if (!bucket) return LoadObject.empty();
-    const items = TaskStore.getItemsInBucket(planId, bucketId);
-    if (items.length === 0) return LoadObject.empty();
-    if (items.length === 1) {
-        return buildSingleTaskRecipeLO(LoadObject.withValue(items[0]))
-            .map(it => ({
-                ...it,
-                name: `${it.name} (${getBucketLabel(bucket)})`,
-            }));
-    }
-    return buildSingleTaskRecipeLO(LoadObject.withValue({
-        name: getBucketLabel(bucket),
-        subtaskIds: items.map(it => it.id),
-    })).map(r => {
-        const idToIdx = new Map();
-        r.subtaskIds.forEach((id, i) => idToIdx.set(id, i));
-        const itToSubIdx = new Map();
-        r.subrecipes.forEach(it => itToSubIdx.set(it, idToIdx.has(it.id) ? idToIdx.get(it.id) : -1));
-        const withPhoto = r.subrecipes.slice().sort((a, b) => {
-            const ai = itToSubIdx.get(a);
-            const bi = itToSubIdx.get(b);
-            if (ai >= 0 && bi < 0) return -1;
-            if (ai < 0 && bi >= 0) return +1;
-            return 0;
-        }).find(it => it.photo);
-        return withPhoto ? {
-            ...r,
-            photo: withPhoto.photo,
-            photoFocus: withPhoto.photoFocus,
-        } : r;
-    });
-};
+import { useLoadedPlan } from "features/RecipeDisplay/hooks/useLoadedPlan";
+import { recipeLoByItemAndBucket } from "features/RecipeDisplay/utils/recipeLoByItemAndBucket";
 
 type Props = RouteComponentProps<{
     pid: string
@@ -64,7 +18,7 @@ const PlannedBucketController: React.FC<Props> = ({ match }) => {
     const pid = parseInt(match.params.pid, 10);
     const bid = parseInt(match.params.bid, 10);
     const lo = useFluxStore(
-        () => buildFullRecipeLO(pid, bid),
+        () => recipeLoByItemAndBucket(pid, bid),
         [
             TaskStore,
             LibraryStore,
