+++ src/features/RecipeDisplay/RecipeController.tsx
@@ -9,47 +9,9 @@ import { useProfileLO } from "providers/Profile";
 import { ScalingProvider } from "util/ScalingContext";
 import LoadingIndicator from "views/common/LoadingIndicator";
 import RecipeDetail from "./components/RecipeDetail";
-import { UserType, } from "global/types/types";
-import { Recipe as RecipeType, } from "features/RecipeDisplay/types";
-import { useGetFullRecipe } from "./hooks/useGetFullRecipe";
-
-export const buildFullRecipeLO = id => {
-    let lo = LibraryStore.getIngredientById(id);
-    if (!lo.hasValue()) return lo;
-
-    const subIds = new Set();
-    const subs: RecipeType[] = [];
-    let loading = false;
-    const prepRecipe = recipe => ({
-        ...recipe,
-        ingredients: (recipe.ingredients || []).map(ref => {
-            if (!ref.ingredientId) return ref;
-            const iLO = LibraryStore.getIngredientById(ref.ingredientId);
-            if (iLO.isLoading()) {
-                loading = true;
-            }
-            if (!iLO.hasValue()) return ref;
-            const ing = iLO.getValueEnforcing();
-            ref = {
-                ...ref,
-                ingredient: ing,
-            };
-            if (ing.type !== "Recipe") return ref;
-            if (subIds.has(ing.id)) return ref;
-            subIds.add(ing.id);
-            subs.push(prepRecipe(ing));
-            return ref;
-        }),
-    });
-    const recipe = prepRecipe(lo.getValueEnforcing());
-    recipe.subrecipes = subs;
-    lo = LoadObject.withValue(recipe);
-    if (loading) {
-        lo = lo.loading();
-    }
-    console.log(lo)
-    return lo;
-};
+import { UserType } from "global/types/types";
+import { Recipe as RecipeType } from "features/RecipeDisplay/types";
+import { recipeLoById } from "features/RecipeDisplay/utils/recipeLoById";
 
 type Props = RouteComponentProps<{
     id: string
@@ -65,13 +27,10 @@ interface State {
 const RecipeController: React.FC<Props> = ({ match }) => {
     const id = parseInt(match.params.id, 10);
 
-    const result = useGetFullRecipe(match.params.id)
-    console.log(result)
-
     const profileLO = useProfileLO();
     const state = useFluxStore(
         () => {
-            const recipeLO = buildFullRecipeLO(id);
+            const recipeLO = recipeLoById(id);
             const state: State = {
                 recipeLO,
                 mine: false,
@@ -97,7 +56,6 @@ const RecipeController: React.FC<Props> = ({ match }) => {
     );
 
     if (state.recipeLO.hasValue()) {
-        console.log(state)
         return <ScalingProvider>
             <RecipeDetail
                 {...state}
