+++ src/features/RecipeDisplay/SharedRecipeController.tsx
@@ -7,12 +7,12 @@ import {
 import { API_BASE_URL } from "constants/index";
 import Dispatcher from "data/dispatcher";
 import LoadObject from "../../util/LoadObject";
-import LoadingIndicator from "../common/LoadingIndicator";
+import LoadingIndicator from "views/common/LoadingIndicator";
 import RecipeDetail from "features/RecipeDisplay/components/RecipeDetail";
 import LibraryActions from "features/RecipeLibrary/data/LibraryActions";
 import { UserType } from "global/types/types";
 import { Maybe } from "graphql/jsutils/Maybe";
-import { buildFullRecipeLO } from "features/RecipeDisplay/RecipeController";
+import { recipeLoById } from "features/RecipeDisplay/utils/recipeLoById";
 
 const axios = BaseAxios.create({
     baseURL: `${API_BASE_URL}/shared/recipe`,
@@ -53,7 +53,7 @@ const DoTheDance: React.FC<MatchParams> = ({
     if (!owner) {
         return <LoadingIndicator />;
     }
-    const recipeLO = buildFullRecipeLO(parseInt(id));
+    const recipeLO = recipeLoById(parseInt(id));
     return <RecipeDetail
         anonymous
         recipeLO={recipeLO}
@@ -67,7 +67,7 @@ interface Props {
     match: match<MatchParams>
 }
 
-const SharedRecipe: React.FC<Props> = props => {
+export const SharedRecipeController: React.FC<Props> = props => {
     const {
         authenticated,
         match: {
@@ -83,5 +83,3 @@ const SharedRecipe: React.FC<Props> = props => {
     }
     return <DoTheDance id={id} {...params} />;
 };
-
-export default SharedRecipe;
