+++ src/features/RecipeDisplay/hooks/useGetFullRecipe.ts
@@ -2,16 +2,14 @@ import { useQuery } from "@apollo/client";
 import { getFullRecipeQuery } from "../data/queries";
 
 export const useGetFullRecipe = (id: string) => {
-    const {loading, error, data } = useQuery(
+    const { loading, error, data } = useQuery(
         getFullRecipeQuery,
-        { variables: { id: id } }
-        )
-
-    console.log(data)
+        { variables: { id: id } },
+    );
 
     return {
         loading,
         error,
-        data
-    }
-}
+        data,
+    };
+};
